Position messages interpolation.
A web.config switch allows to enable position message interpolation
active during position messages storage. Interpolation aims at limiting
storage space as low as possible while keeping useful informationi. It
is expecially useful when keep-alive messages are used to track a
stationary vehicle. Keep-alive messages allow to know that the GPS
tracking is fully working, though the position messages don't actually update
vehicle position at all. In this case, it has sense to keep only the first and the last
arrived messages.

In order to reach this objective, interpolation acts as follows: on position
message arrival the last 3 messages for the same vehicle are analyzed (the one just arrived and the
next two stored in the database); in case such three messages actually exist,
and they are "close enough" each other, the middle message (i.e. the
second one) is deleted from the database.

The last stored message carries information about interpolation action:

* how many messages it interpolates;
* how long it has been interpolating;
* datetime of the closest interpolated message.

"Close enough" above depends on the threshold available in the
web.config, expressed in meters.

In case the last stored message does not interpolate any message, the interpolated
data is null.

As an example, in case 10 messages indicating the same position for the
same vehicle arrive, one per minute, only the first and the last are stored, as shown
below:

{
    "_id" : "5aa6e2f24eb6d90fd431b73d",
    "codiceMezzo" : "M.1136",
    "localizzazione" : {
        "type" : "Point",
        "coordinates" : [
            1.4,
            1.5
        ]
    },
    "istanteAcquisizione" : ISODate("2018-03-12T20:55:52.000Z"),
    "ultimo" : true,
    "interpolationData" : {
        "length_sec" : 540,
        "messages" : 9,
        "lastMsgTime" : ISODate("2018-03-12T20:54:51.000Z")
    }
},
{
    "_id" : "5aa6e2f24eb6d90fd431b700",
    "codiceMezzo" : "M.1136",
    "localizzazione" : {
        "type" : "Point",
        "coordinates" : [
            1.4,
            1.5
        ]
    },
    "istanteAcquisizione" : ISODate("2018-03-12T20:46:52.000Z"),
    "ultimo" : false,
    "interpolationData" : null
}