Lock on vehicleCode upon position message insertion
Upon position message insertion, a lock is acquired on vehicleCode.
Thus, two concurrent thread simultaneously inserting on the same vehicle
(which is unlikely but not impossible) cannot cause a race condition.

This update requires web.config to be extended with this two new
parameters:

<!-- vehicle lock on insertion parameters - start -->

    <!-- number of attempts to acquire the lock, after which the insertion gives up. -->
    <add key="numberOfRetries" value="5" />
    <!-- interval between attempts to acquire lock -->
    <add key="retriesInterval_msec" value="5000" />

<!-- vehicle lock on insertion parameters - end -->

Locking is managed through a new collection called vehicleLocks. This
collection is almost always empty, except during an insertion. The
inserting thread atomically create a document within this collection,
having the _id field equal to the vehicle code to be inserted. One more
such record cannot be isimultaneously inserted by another thread for the
same vehicle since this would violate the _id unicity.

vehicleLocks collection has an expiring index on the ts (insertion time)
field. This prevents an everlasting lock in case of application failure
during an insertion (see
https://docs.mongodb.com/manual/tutorial/expire-data/).