IP-based authorization to post position messages
Posting position messages can be subject to client authorization, based on IP source address. The web.config allows to enable/disable IP-based authorization through a flag and to specify the list of allowed IP and/or networks (e.g. 127.0.0.1, 8.8.8.8, 10.0.0.0/24, 172.16.0.0/16).

The web.config section to allow IP-based authorization is structured as
follows:

<!-- authorization - start -->

    <!-- enables IP based authorization for posting position messages -->
    <add key="ipAuthorizationEnabled" value="false" />
    <!-- list of networks authorized to post position messages (e.g. 127.0.0.1, 8.8.8.8, 10.0.0.0/24, 172.16.0.0/16) -->
    <add key="authorizedIpSources" value="127.0.0.1, 10.0.0.0/24" />

<!-- authorization - end -->