http file to test some actions
http file is used to encode some action requests to be executed on the
backend. This file is convieniently read by the REST client extension in
Visual Studio Code. This extension allows to trigger actions and reading
returned results in a straighforward way. http files can be easily
versioned, too.