Jitter on message insertion
When a message insertion occurs, the application tries to create a database lock on the vehicle, in order to grant a correct transaction. When the lock creation fails, the application retries a certain number of times (as specified in the configuration file). With this commit, the time elapsed between retrials is composed also of a random contribution, in order to prevent that many insertions retry at exactly the same time.