Connection string no longer parsed
Custom connection string parsing did not work and was not cluster compliant. In order to let MongoDB parse the connection string, and in order to enable mongodb+srv//: connection strings formats, the connection string is no longer parsed and the client is directly created from the connection string contained within the Web.Config.

WARNING: due to this change, mongodb queries are no longer logged. A fix is needed for this.

Apart from this major change, Owin and compiler libraries have been updated.